add test, use cell.point in iris Constraint definition so that it does not take into account the eventual bounds
