+++ tests/conftest.py
@@ -13,6 +13,7 @@ pytest_plugins = [
     "tests.fixtures.aeronet",
     "tests.fixtures.stations",
     "tests.fixtures.collocated_data",
+    "tests.fixtures.dummy_model_data",
     "tests.fixtures.aeroval.config",
     "tests.fixtures.cams2_83.config",
     "tests.fixtures.pyaro",
